alert add plant

+++ server/app/Http/Controllers/Admin/PlantsController.php
@@ -25,48 +25,70 @@ class PlantsController extends Controller
       $plant =Plant::create($request->all());
        return redirect()->route('admin.plants.index');
     }
-    public function edit($id){
-        $plant= Plant::find($id);
-        $lights= Light::pluck('name');
-        $environments= Environment::pluck('name');
-        return view('admin.plants.edit', compact('plant','lights','environments'));
-        //return view('admin.plants.edit', compact('plant','id'));
-    }
-    
     public function show(Plant $plant)
     {
-        return view('admin.plants.show', compact('plant'));
+       // return view('admin.plants.show', compact('plant'));
     }
-  /*  public function update(UpdateRequest $request, Plant $plant)
-    {   //falta colocar require sino no no funciona
-        $request -> validated([
-            'name' => 'require',
-            'environment_id' => 'required',
-            'light_id' => 'required',
-            'date' => 'required',
-            'description' => 'required',
-            'image' => 'sometimes|required:NULL',
-            'user_id' => auth()->id(),
-        ]);
-        $plant->update($request->all());
-        return redirect()->route('admin.plants.show',$plant)->with('info','se actualizo con exit');
-    }*/
-    public function  update(Request $request, Plant $plant){
-        $plant->name = $request-> name;
-        $plant->environment_id = $request-> enviroment_id;
-        $plant->light_id = $request-> light_id;
-        $plant-> date= $request-> date;
-        $plant-> description= $request-> description;
-        $plant-> image= $request->image ;
+    public function edit($id){
+        $plant= Plant::find($id);
+       // $lights= Light::pluck('name');
+       // $environments= Environment::pluck('name');
+       // return view('admin.plants.edit', compact('plant','lights','environments'));
+        return view('admin.plants.edit', ['plant'=>$plant]);
+    }
+    
+    public function update(Request $request, string $id)
+    {
+        try {
+            
+            $request->validate([
+        'name' => 'nullable|string',
+        'environment_id' => 'nullable|integer',
+        'light_id' => 'nullable|integer',
+        'date' => 'nullable|date', 
+        'description' => 'nullable|string',
+        'image' => 'nullable|string',
+        'user_id' => 'nullable|integer',
+    ]);
+     $plant = Plant::findOrFail($id);
+       // $user->name = $request->input('name');
+
+        $plant->name = $request->input('name');
+        $plant->environment_id = $request->input('environment_id');
+        $plant->light_id = $request->input('light_id');
+        $plant->date= $request->input('date');
+        $plant->description= $request->input('description');
+        $plant->image= $request->input('image') ;
+        $plant->user_id=$request->input('user_id');
+
+
         $plant->save();
-        return redirect()->route('admin.plants.show', $plant);
+       // return redirect()->route('admin.plants.show', $plant);
+       return response()->json(['success' => true, 'message' => 'Planta actualizado correctamente']);
+    } catch (\Throwable $th) {
+        return response()->json(['success' => false, 'message' => 'La planta no se pudo actualizar']);
     }
 
-    public function destroy(Plant $plant) //no elimina
+}
+
+
+   // public function destroy(Plant $plant) //no elimina
+    //{
+     //   $plant->delete();
+     //   return redirect()->route('admin.plants.index',$plant);//->with('info',' se elimino con exit');
+    //} 
+    public function destroy(string $id)
     {
+    $plant = Plant::find($id);
+
+    if ($plant) {
         $plant->delete();
-        return redirect()->route('admin.plants.index',$plant);//->with('info',' se elimino con exit');
-    } 
+        return response()->json(['success' => true, 'message' => 'Usuario eliminado correctamente']);
+    } else {
+        return response()->json(['success' => false, 'message' => 'El usuario no existe o no se pudo eliminar']);
+    }
+    }
+
 
    
 }
